+++ test/Main.hs
@@ -28,6 +28,7 @@ import qualified Data.Bits as Bits
 import qualified Data.Bytes.Parser as P
 import qualified Data.Bytes.Parser.Ascii as Ascii
 import qualified Data.Bytes.Parser.Latin as Latin
+import qualified Data.Bytes.Parser.Leb128 as Leb128
 import qualified Data.Bytes.Parser.BigEndian as BigEndian
 import qualified Data.Bytes.Parser.LittleEndian as LittleEndian
 import qualified Data.Primitive as PM
@@ -357,6 +358,26 @@ tests = testGroup "Parser"
         P.parseBytes (Latin.hexFixedWord64 ()) (bytes "ABCD01235678BCDE")
         @=? P.Success
         (Slice 17 0 0xABCD01235678BCDE)
+    ]
+  , testGroup "leb128-w32"
+    [ testCase "A" $
+        P.Success (Slice 2 0 0x7E)
+        @=?
+        P.parseBytes (Leb128.word32 ()) (bytes "\x7E")
+    , testCase "B" $
+        P.Success (Slice 5 0 0x200000)
+        @=?
+        P.parseBytes (Leb128.word32 ()) (bytes "\x81\x80\x80\x00")
+    ]
+  , testGroup "leb128-w16"
+    [ testCase "A" $
+        P.Failure ()
+        @=?
+        P.parseBytes (Leb128.word16 ()) (bytes "\x84\x80\x00")
+    , testCase "B" $
+        P.Success (Slice 4 0 0xFFFF)
+        @=?
+        P.parseBytes (Leb128.word16 ()) (bytes "\x83\xFF\x7F")
     ]
   ]